co-locate specs next to their files
