+++ webpack/apps/original-prototype/components/Autosuggest/FairAutosuggest.spec.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import renderer from 'react-test-renderer'
+import { FairAutosuggest } from './FairAutosuggest'
+
+const mockHandler = jest.fn()
+
+describe('FairAutosuggest', () => {
+  it('renders correctly', () => {
+    const rendered = renderer.create(<FairAutosuggest onSelectFair={mockHandler} />)
+    const tree = rendered.toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+})
