+++ webpack/apps/original-prototype/components/Autosuggest/GeneAutosuggest.spec.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import renderer from 'react-test-renderer'
+import { GeneAutosuggest } from './GeneAutosuggest'
+
+const mockHandler = jest.fn()
+
+describe('GeneAutosuggest', () => {
+  it('renders correctly', () => {
+    const rendered = renderer.create(<GeneAutosuggest onSelectGene={mockHandler} />)
+    const tree = rendered.toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+})
+
